requirements.txt
cvlib
opencv-python
ultralytics
matplotlib
replicate
python-dotenv
gradio